Fix all uses of assert
Assert is not run in release builds, as if the whole line is deleted. So
anything with side effects in an assert only runs in the debug build,
includind all the signal connections and tweening I had before. Moving
all the actual code out of the assert and asserting on the return value
fixes this.