Fix some lag spikes caused by shader compilation
Mostly noticable on Firefox.

Shader compilation seems to only happen on first render, so added
some planes and dummy props behind the wall in the main menu with
all the materials in the game. There's no visibility culling so
as long as they're in the view frustum they'll get rendered and trigger
the shader compilation right as the menu loads.